% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_expProfiles.R
\name{add_padding}
\alias{add_padding}
\title{Adds padding to a exposure profile}
\usage{
add_padding(profile, dt, windowLength)
}
\arguments{
\item{profile}{FOCUS profile before any padding}

\item{dt}{the timesteps at which to add to the time and conc vectors, (days), smaller means more accurate}

\item{windowLength}{the length of each time window}
}
\value{
an exposure profile with padding
}
\description{
Adds padding of zero concentration to the start and end of the FOCUS profile equal to the windowLength
minus 1 dt step. Thus the first and last windows have minimum (but not zero) time with non-zero exposure.
}
