% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_toxswa_profiles.R
\name{check_if_TOXSWA_version}
\alias{check_if_TOXSWA_version}
\title{Checks if the toxswa version is as specified}
\usage{
check_if_TOXSWA_version(ln, supported_version = c("4", "5.5.3"))
}
\arguments{
\item{ln}{a vector of strings}

\item{supported_version}{a string of the version that is looked for}
}
\value{
TRUE if version numbers match
}
\description{
Checks if the toxswa version is as specified
}
\examples{
ln <- c("* ------------------------------------------------------------------------------",
"* Results from the TOXSWA model  (c) Wageningen University & Research",
"* FOCUS  TOXSWA version   : 5.5.3",
"* TOXSWA model version    : 3.3.6",
"* TOXSWA created on       : 17-Dec-2017")
deeptox:::check_if_TOXSWA_version(ln, supported_version = "5.5.3")

 ln2 <- c("TOXSWA REPORT: Header",
 "* Results from the TOXSWA model  (c) Alterra",
 "* FOCUS  TOXSWA version   : 4",
 "* TOXSWA model version    : 3.3.4",
 "* TOXSWA created on       : 02-Apr-2015")
deeptox:::check_if_TOXSWA_version(ln2)
}
