% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deeptox_ExampleDataset}
\alias{deeptox_ExampleDataset}
\title{Complete dataset of finished DEB TKTD predictions}
\format{
A list with 5 elements:
\describe{
  \item{expProfiles}{the input exposure profiles}
  \item{parameters}{the parameter set used}
  \item{profileResults}{the results generated}
  \item{summaryResults}{the summarized results}
  \item{plotResults}{the generated plots}
}
}
\usage{
deeptox_ExampleDataset
}
\description{
The dataset is a pregenerated complete dataset with default input parameter set as 
calibrated for Daphnia magna and an exposure profile set for different exposure 
scenarios in apples and cereals. The dataset includes all generated results with
tables and figures.
}
\keyword{datasets}
