% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_parameters.R
\name{exportParsAndSettings}
\alias{exportParsAndSettings}
\title{Export parameters and setting}
\usage{
exportParsAndSettings(parsAndSettings, outfile, parorder = NA)
}
\arguments{
\item{parsAndSettings}{the parameters and settings to be exported}

\item{outfile}{the path to the file to which the parameters and settings are written}

\item{parorder}{the order of the parameters and settings}
}
\value{
write table to 'outfile' file
}
\description{
Export parameters and setting
}
