% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{flatten_list}
\alias{flatten_list}
\title{Flatten a list to first layer}
\usage{
flatten_list(lst)
}
\arguments{
\item{lst}{the input list}
}
\value{
a list
}
\description{
Flatten a list to first layer
}
\examples{
iris_list <- split(iris, iris$Species)
nested_list <- list(iris_list[1], list(iris_list[2], iris_list[3]))
nested_list
flatten_list(nested_list)
}
