% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_parameters.R
\name{importParsAndSettings}
\alias{importParsAndSettings}
\title{Import parameters and setting}
\usage{
importParsAndSettings(parameters_rv, pathname)
}
\arguments{
\item{parameters_rv}{reactive values object to which the contents of the file are stored}

\item{pathname}{file whose contents are imported}
}
\value{
a reactive values object
}
\description{
Import parameters and setting
}
