% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{make.names2}
\alias{make.names2}
\title{Make Syntactically Valid Names}
\usage{
make.names2(names_, replace_char = ".", ...)
}
\arguments{
\item{names_}{character vector to be coerced to syntactically valid names. This is coerced to character if necessary.}

\item{replace_char}{the character to be used to replace the default "." of make.names.}

\item{...}{arguments passed to make.names}
}
\value{
A character vector of same length as names with each changed to a syntactically valid name, in the current locale's encoding.
}
\description{
Make syntactically valid names out of character vectors.
Based on base::make.names function. Added functionality to include the
}
\examples{
strings <- c("a and b", "a-and-b")
# same results as make.names
make.names2(strings, unique = TRUE)
make.names(strings, unique = TRUE)

# replace character "_" instead of "."
make.names2(strings, replace_char = "_", unique = TRUE)
}
