% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_windows.R
\name{moving_time_windows}
\alias{moving_time_windows}
\title{Apply moving time windows approach to exposure assessment}
\usage{
moving_time_windows(
  windows,
  parms,
  y0,
  cutOff,
  ECpct,
  tol = 0.00001,
  detail,
  passFail,
  progressBar = "txt",
  scenName = ""
)
}
\arguments{
\item{windows}{list of lists - vectors for times and concentrations}

\item{cutOff}{safety cutoff that is good enough for us to move on}

\item{ECpct}{what percentage of effects do we want to see? Probably 10}

\item{tol}{tolerance around ECpct acceptable from the algorithm}

\item{detail}{Boolean, do we want to only get pass or fail results at passFail value?}

\item{passFail}{the passFail AF - usually 10.}

\item{progressBar}{defines the type of progress bar either "txt" for a text based progress bar or "shiny" for a shiny progress bar}
}
\value{
results of the EPx analysis for the moving time windows
}
\description{
Assesses the effects of exposure in each time window at the cutOff and passFail values.
If detail is TRUE, it also calculates the exact values.
}
