% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_windows.R
\name{organise_windows}
\alias{organise_windows}
\title{Organise windows}
\usage{
organise_windows(profile, windowLength, interval)
}
\arguments{
\item{profile}{2 column data set: [time, conc]}

\item{windowLength}{duration (in days) of each window}

\item{interval}{how far apart each window is.}
}
\value{
time windows
}
\description{
Takes the given profile and returns a list (of lists) with time and concentration vectors for all time windows
}
