% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_results.R
\name{plotEPx}
\alias{plotEPx}
\title{Plot EPx}
\usage{
plotEPx(
  sT,
  EP_V,
  interval,
  tmax,
  windowLength = 21,
  passFail = 10,
  cutOff = NA,
  xunit = NA
)
}
\arguments{
\item{sT}{startTimes of all windows}

\item{EP_V}{EPx values of all windows}

\item{interval}{length of time (in days) between starts of successive windows}

\item{tmax}{end time of the profile}

\item{windowLength}{How long (days) is the time window}

\item{passFail}{chronic AF value used in all risk assessments for pass/fail.}

\item{cutOff}{the cutOff value is used to determine the y-axis maximum}
}
\value{
a ggplot object of the EPx
}
\description{
function plots the EPx multiplier curve. In cases where thinning has been used multiple
 separate lines are plotted.
}
