% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_expProfiles.R
\name{plotExpProfiles}
\alias{plotExpProfiles}
\title{Plot exposure profile}
\usage{
plotExpProfiles(data, x = "V1", y = "V2", title = NULL, xunit = NA, yunit = NA)
}
\arguments{
\item{data}{the exposure profile data loaded with 'readExpProfiles'}

\item{x}{the data column plotted on the x-axis}

\item{y}{the data column plotted on the y-axis}

\item{title}{the plot title}

\item{xlab}{the label of the x-axis}

\item{ylab}{the label of the y-axis}
}
\value{
a ggplot object of exposure profiles
}
\description{
Plot exposure profile
}
