% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_results.R
\name{plotFailures_rect}
\alias{plotFailures_rect}
\title{Plot failures_rect}
\usage{
plotFailures_rect(
  profile,
  sT,
  EP_V,
  interval,
  tmax,
  windowLength = 21,
  passFail = 10,
  yunit = NA,
  xunit = NA
)
}
\arguments{
\item{profile}{time and concentration vectors for the whole profile plus padding}

\item{sT}{start times of all windows}

\item{EP_V}{EPx results for all windows}

\item{interval}{length of time (in days) between starts of successive windows}

\item{tmax}{end time of the profile}

\item{windowLength}{How long (days) is the time window}

\item{passFail}{chronic AF value used in all risk assessments for pass/fail.}
}
\value{
a ggplot object of the failures
}
\description{
function plots the EPx multiplier curve. In cases where thinning has been used multiple
 separate lines are plotted.
}
