% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_toxswa_profiles.R
\name{pre_check_toxswa_files}
\alias{pre_check_toxswa_files}
\title{Check if the selected files are in a supported TOXSWA format}
\usage{
pre_check_toxswa_files(file_names, profileNames)
}
\arguments{
\item{file_names}{the name of the file that is going to be imported}

\item{profileNames}{the short names for the profiles}
}
\value{
an HTML string indicating which file
}
\description{
Check if the selected files are in a supported TOXSWA format
}
