% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_results.R
\name{profileSummary}
\alias{profileSummary}
\title{Summarize profiles}
\usage{
profileSummary(profileRes, passFail, roundEPx = NA)
}
\arguments{
\item{profileRes}{the profile results}

\item{passFail}{the threshold at which pass/fail are assessed}

\item{roundEPx}{the number of digits to which the EPx should be rounded; NA: no rounding}
}
\value{
summary for a single profile
}
\description{
Summarize profiles
}
