% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_expProfiles.R
\name{readExpProfiles}
\alias{readExpProfiles}
\title{Read exposure profiles}
\usage{
readExpProfiles(pathtofiles, profileNames = NA)
}
\arguments{
\item{pathtofiles}{}

\item{profileNames}{}
}
\value{
list of exposure profiles
}
\description{
Read several exposure profiles at their corresponding pathes at 'pathtofiles'
}
