% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_expProfiles.R
\name{readSingleExpProfile}
\alias{readSingleExpProfile}
\title{Read an exposure profile}
\usage{
readSingleExpProfile(pathtofile)
}
\arguments{
\item{pathtofile}{the path where the exposure profile is found}
}
\value{
the exposure profile
}
\description{
Read a single exposure profile at 'pathtofile'
}
