% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_toxswa_profiles.R
\name{readToxswaExpProfiles}
\alias{readToxswaExpProfiles}
\title{Function to read TOXSWA exposure profiles from out-files}
\usage{
readToxswaExpProfiles(pathtofiles, profileNames = NA)
}
\arguments{
\item{pathtofiles}{path to toxswa files that should be read}

\item{profileNames}{a vector of names to be used for the loaded toxswa files. if "NA" the filename without extension will be used}
}
\value{
list of exposure profiles each with two columns (time and concentration)
}
\description{
Read several expsoure profiles at their corresponding pathes at 'pathtofiles'
}
