% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_toxswa_profiles.R
\name{split_toxswa_exposure_profile}
\alias{split_toxswa_exposure_profile}
\title{Split a profile into a list if more then one active substances was detected}
\usage{
split_toxswa_exposure_profile(profile, runid = NA)
}
\arguments{
\item{profile}{the toxswa profile as a tibble}
}
\value{
a tibble or a list of tibbles if more than one active substance was identified
}
\description{
Split a profile into a list if more then one active substances was detected
}
\examples{
dat <- data.frame(RunID = 123, 
  Layer_Specifier = rep(c("ConLiqWatLay_ai1","ConLiqWatLay_ai2"),10),
  Concentration = rlnorm(20,0,1)
  )
  deeptox:::split_toxswa_exposure_profile(dat)
}
