% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_results.R
\name{summaryForSingleProfile}
\alias{summaryForSingleProfile}
\title{Summarize single profile}
\usage{
summaryForSingleProfile(
  profileName,
  windows,
  windows_proc,
  thisProfile,
  passFail,
  roundEPx = NA
)
}
\arguments{
\item{profileName}{the name of the profile to analyse}

\item{windows}{all windows available for all profiles}

\item{windows_proc}{processed windows for all profiles}

\item{thisProfile}{the profile results}

\item{passFail}{the threshold at which pass/fail are assessed}

\item{roundEPx}{the number of digits to which the EPx should be rounded; NA: no rounding}
}
\value{
summary table for a single profile
}
\description{
Summarize single profile
}
