% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_windows.R
\name{thinWindows}
\alias{thinWindows}
\title{Thin windows}
\usage{
thinWindows(windows)
}
\arguments{
\item{windows}{2dim array (list of lists), times concs}
}
\value{
time windows
}
\description{
thin out the list of windows to only those which might be worst case scenario
}
