% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_results.R
\name{VTTsummary}
\alias{VTTsummary}
\title{Summarize VTTs}
\usage{
VTTsummary(profileName, windows, windows_proc)
}
\arguments{
\item{profileName}{the name of the profile to analyse}

\item{windows}{all windows available for all profiles}

\item{windows_proc}{processed windows for all profiles}
}
\value{
a summary of the VTT summary
}
\description{
The function summarized the virtual toxicity tests (VTTs)
}
